windows / layouts/ rowColumnLayout
excerpt from the Maya commands docs

```
This command creates a rowColumn layout. A rowColumn layout positions children in either a row or column format. A column layout, specified with the -nc/numberOfColumns flag, allows you set text alignment, attachments and offsets for each column in the layout. Every member of a column will have the same alignment, attachment and offsets. Likewise the row format, specified by the -nr/numberOfRows flag, allows setting of these attributes for each row in the layout. Every member of a row will have the same attributes. The layout must be either a row or column format. This layout does not support both, or the specification of attributes on an individual child basis.

Some flags only make sense for one of either the row format or the column format. For example the -rh/rowHeight flag can only be specified in row format. In column format the row height is determined by the tallest child in the row, plus offsets. 
```